// report-ng/app-react/src/model/report-model/framework_pb.ts
export const FailureCorridorValue = { FCV_NOT_SET: 0, FCV_HIGH: 1, FCV_MID: 2, FCV_LOW: 3 } as const;

export type FailureCorridorValue = typeof FailureCorridorValue[keyof typeof FailureCorridorValue];

export namespace FailureCorridorValue {
  export type FCV_NOT_SET = typeof FailureCorridorValue.FCV_NOT_SET;
  export type FCV_HIGH = typeof FailureCorridorValue.FCV_HIGH;
  export type FCV_MID = typeof FailureCorridorValue.FCV_MID;
  export type FCV_LOW = typeof FailureCorridorValue.FCV_LOW;
}

export const MethodType = { MT_NOT_SET: 0, TEST_METHOD: 1, CONFIGURATION_METHOD: 2 } as const;

export type MethodType = typeof MethodType[keyof typeof MethodType];

export namespace MethodType {
  export type MT_NOT_SET = typeof MethodType.MT_NOT_SET;
  export type TEST_METHOD = typeof MethodType.TEST_METHOD;
  export type CONFIGURATION_METHOD = typeof MethodType.CONFIGURATION_METHOD;
}

/** All status elements for an actual executed test method */
export const ResultStatusType = {
  RST_NOT_SET: 0,
  /** NO_RUN - default status when entity is spawned, basically an illegal end status */
  NO_RUN: 1,
  /**
   * INFO - info status, not representative
   *
   * @deprecated
   */
  INFO: 2,
  /** SKIPPED - skipped */
  SKIPPED: 3,
  /** PASSED - passed status, without any issue */
  PASSED: 4,
  /**
   * MINOR - passed, but with minor, non-blocking issues
   *
   * @deprecated
   */
  MINOR: 5,
  /** FAILED - failed status, with hard failing issues */
  FAILED: 7,
  /**
   * FAILED_MINOR - failed with additional minor issues
   *
   * @deprecated
   */
  FAILED_MINOR: 8,
  /** FAILED_RETRIED - failed and retry is triggered */
  FAILED_RETRIED: 9,
  /** FAILED_EXPECTED - entity was expected to fail (with or without an explicit issue), it may or may not be representative */
  FAILED_EXPECTED: 10,
  /** PASSED_RETRY - passed after a retry */
  PASSED_RETRY: 11,
  /**
   * MINOR_RETRY - minor after a retry
   *
   * @deprecated
   */
  MINOR_RETRY: 12,
  /** REPAIRED - passed with fail annotation */
  REPAIRED: 13,
} as const;

export type ResultStatusType = typeof ResultStatusType[keyof typeof ResultStatusType];
// ...
  export type RST_NOT_SET = typeof ResultStatusType.RST_NOT_SET;
  export type NO_RUN = typeof ResultStatusType.NO_RUN;
  export type INFO = typeof ResultStatusType.INFO;
  export type SKIPPED = typeof ResultStatusType.SKIPPED;
  export type PASSED = typeof ResultStatusType.PASSED;
  export type MINOR = typeof ResultStatusType.MINOR;
// ...
export interface MethodContext {
  contextValues?: ContextValues | undefined;
  methodType?:
    | MethodType
    | undefined;
  /**
   * repeated string parameters = 8 [deprecated = true];
   *    repeated string method_tags = 9 [deprecated = true];
   */
  retryNumber?: number | undefined;
  methodRunIndex?: number | undefined;
  threadName?:
    | string
    | undefined;
  /** TestStep failed_step = 13; */
  failureCorridorValue?: FailureCorridorValue | undefined;
  classContextId?:
    | string
    | undefined;
  /** @deprecated */
  executionContextId?:
    | string
    | undefined;
  /**
   * repeated ErrorContext non_functional_infos = 17 [deprecated = true];
   *    repeated ErrorContext collected_assertions = 18 [deprecated = true];
   *
   * @deprecated
   */
  infos?:
    | string[]
    | undefined;
  /** @deprecated */
  priorityMessage?: string | undefined;
  relatedMethodContextIds?: string[] | undefined;
  dependsOnMethodContextIds?:
    | string[]
    | undefined;
  /** @deprecated */
  errorContext?: ErrorContext | undefined;
  testSteps?:
    | TestStep[]
    | undefined;
  /** @deprecated */
  testContextId?:
    | string
    | undefined;
  /** @deprecated */
  suiteContextId?: string | undefined;
  sessionContextIds?:
    | string[]
    | undefined;
  /**
   * repeated string video_ids = 30  [deprecated = true];
   *    repeated string screenshot_ids = 31 [deprecated = true];
   *    string custom_context_json = 32 [deprecated = true];
   */
  failedStepIndex?: number | undefined;
  resultStatus?: ResultStatusType | undefined;
  parameters?: { [key: string]: string } | undefined;
  customContexts?: { [key: string]: string } | undefined;
  annotations?:
    | { [key: string]: string }
    | undefined;
  /** A custom generated test name (e.a. cucumber scenario) */
  testName?: string | undefined;
  layoutCheckContext?: LayoutCheckContext[] | undefined;
}

export interface MethodContext_ParametersEntry {
  key: string;
  value: string;
}

export interface MethodContext_CustomContextsEntry {
  key: string;
  value: string;
}

export interface MethodContext_AnnotationsEntry {
  key: string;
  value: string;
}

export interface ContextValues {
  id?: string | undefined;
  created?: number | undefined;
  name?: string | undefined;
  startTime?:
    | number
    | undefined;
  /**
   * string swi = 6 [deprecated = true];
   *    ResultStatusType result_status = 7 [deprecated = true];
   *    ExecStatusType exec_status = 8 [deprecated = true];
   */
  endTime?: number | undefined;
}

export interface TestStep {
  name?:
    | string
    | undefined;
  /** string id = 2; */
  actions?: TestStepAction[] | undefined;
}
// ...
export interface ErrorContext {
  /**
   * string readable_error_message = 1 [deprecated = true];
   *    string additional_error_message = 2 [deprecated = true];
   *    StackTrace stack_trace = 3  [deprecated = true];
   *    string error_fingerprint = 6 [deprecated = true];
   */
  scriptSource?:
    | ScriptSource
    | undefined;
  /**
   * ScriptSource execution_object_source = 8 [deprecated = true];
   *
   * @deprecated
   */
  ticketId?:
    | string
    | undefined;
  /** @deprecated */
  description?: string | undefined;
  stackTrace?: StackTraceCause[] | undefined;
  optional?: boolean | undefined;
  id?: string | undefined;
}
// ...
export interface LayoutCheckContext {
  image?: string | undefined;
  distance?: number | undefined;
  expectedScreenshotId?: string | undefined;
  actualScreenshotId?: string | undefined;
  distanceScreenshotId?: string | undefined;
  errorContextId?: string | undefined;
}
// ...
export interface StackTraceCause {
  className?: string | undefined;
  message?:
    | string
    | undefined;
  /** StackTraceCause cause = 4 [deprecated = true]; */
  stackTraceElements?: string[] | undefined;
}

export interface ScriptSource {
  fileName?: string | undefined;
  methodName?: string | undefined;
  lines?: ScriptSourceLine[] | undefined;
  mark?: number | undefined;
}
// ...
export const ErrorContext: MessageFns<ErrorContext> = {
  encode(message: ErrorContext, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.scriptSource !== undefined) {
      ScriptSource.encode(message.scriptSource, writer.uint32(58).fork()).join();
    }
    if (message.ticketId !== undefined && message.ticketId !== "") {
      writer.uint32(74).string(message.ticketId);
    }
    if (message.description !== undefined && message.description !== "") {
      writer.uint32(82).string(message.description);
    }
    if (message.stackTrace !== undefined && message.stackTrace.length !== 0) {
      for (const v of message.stackTrace) {
        StackTraceCause.encode(v!, writer.uint32(90).fork()).join();
      }
    }
    if (message.optional !== undefined && message.optional !== false) {
      writer.uint32(96).bool(message.optional);
    }
    if (message.id !== undefined && message.id !== "") {
      writer.uint32(106).string(message.id);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ErrorContext {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseErrorContext();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 7: {
          if (tag !== 58) {
            break;
          }

          message.scriptSource = ScriptSource.decode(reader, reader.uint32());
          continue;
        }
        case 9: {
          if (tag !== 74) {
            break;
          }

          message.ticketId = reader.string();
          continue;
        }
        case 10: {
          if (tag !== 82) {
            break;
          }

          message.description = reader.string();
          continue;
        }
        case 11: {
          if (tag !== 90) {
            break;
          }

          const el = StackTraceCause.decode(reader, reader.uint32());
          if (el !== undefined) {
            message.stackTrace!.push(el);
          }
          continue;
        }
        case 12: {
          if (tag !== 96) {
            break;
          }

          message.optional = reader.bool();
          continue;
        }
        case 13: {
          if (tag !== 106) {
            break;
          }

          message.id = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },
};
// ...
export const StackTraceCause: MessageFns<StackTraceCause> = {
  encode(message: StackTraceCause, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.className !== undefined && message.className !== "") {
      writer.uint32(10).string(message.className);
    }
    if (message.message !== undefined && message.message !== "") {
      writer.uint32(18).string(message.message);
    }
    if (message.stackTraceElements !== undefined && message.stackTraceElements.length !== 0) {
      for (const v of message.stackTraceElements) {
        writer.uint32(26).string(v!);
      }
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): StackTraceCause {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseStackTraceCause();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.className = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.message = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          const el = reader.string();
          if (el !== undefined) {
            message.stackTraceElements!.push(el);
          }
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },
};

function createBaseScriptSource(): ScriptSource {
  return { fileName: "", methodName: "", lines: [], mark: 0 };
}

export const ScriptSource: MessageFns<ScriptSource> = {
  encode(message: ScriptSource, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.fileName !== undefined && message.fileName !== "") {
      writer.uint32(10).string(message.fileName);
    }
    if (message.methodName !== undefined && message.methodName !== "") {
      writer.uint32(18).string(message.methodName);
    }
    if (message.lines !== undefined && message.lines.length !== 0) {
      for (const v of message.lines) {
        ScriptSourceLine.encode(v!, writer.uint32(26).fork()).join();
      }
    }
    if (message.mark !== undefined && message.mark !== 0) {
      writer.uint32(32).int32(message.mark);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ScriptSource {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseScriptSource();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.fileName = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.methodName = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          const el = ScriptSourceLine.decode(reader, reader.uint32());
          if (el !== undefined) {
            message.lines!.push(el);
          }
          continue;
        }
        case 4: {
          if (tag !== 32) {
            break;
          }

          message.mark = reader.int32();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },
};
// ...
interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
}